feat(weather): 新增区域信息管理功能
- 添加 RegionalInfo 相关的 Controller、Service、Mapper及对应的 VO/BO 类
- 实现区域信息的增删改查功能
- 添加数据校验和分页查询支持

+++ modules/weather/src/main/java/com/cdzeroly/weather/domain/vo/GroundStationVo.java
@@ -55,13 +55,13 @@ public class GroundStationVo implements Serializable {
      * 经度
      */
     @ExcelProperty(value = "经度")
-    private Long longitude;
+    private Double longitude;
 
     /**
      * 纬度
      */
     @ExcelProperty(value = "纬度")
-    private Long latitude;
+    private Double latitude;
 
     /**
      * 地理坐标
